feat: create inheritance between Conteudo with Curso and Mentoria
