+++ src/br/com/dio/desafio/dominio/Curso.java
@@ -1,26 +1,8 @@
 package br.com.dio.desafio.dominio;
 
-public class Curso {
-    private String titulo;
-    private String descricao;
+public class Curso extends Conteudo{
+
     private int cargaHoraria;
-
-    public String getTitulo() {
-        return titulo;
-    }
-
-    public void setTitulo(String titulo) {
-        this.titulo = titulo;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-
-    public void setDescricao(String descricao) {
-        this.descricao = descricao;
-    }
-
     public int getCargaHoraria() {
         return cargaHoraria;
     }
@@ -30,11 +12,20 @@ public class Curso {
     }
 
     @Override
+    public double calcularXp() {
+        return XP_PADRAO * cargaHoraria;
+    }
+
+    public Curso() {
+    }
+
+    @Override
     public String toString() {
         return "Curso{" +
-                "titulo='" + titulo + '\'' +
-                ", descricao='" + descricao + '\'' +
-                ", cargaHoraria=" + cargaHoraria +
+                "titulo='" + getTitulo()+ '\'' +
+                ", descricao='" + getDescricao()+ '\'' +
+                ", cargaHoraria=" +  cargaHoraria +
                 '}';
     }
+
 }
